+++ src/main/java/com/atlauncher/gui/dialogs/ModrinthVersionSelectorDialog.java
@@ -194,7 +194,7 @@ public class ModrinthVersionSelectorDialog extends JDialog {
             } else if (App.settings.addModRestriction == AddModRestriction.LAX) {
                 try {
                     List<String> minecraftVersionsToSearch = MinecraftManager
-                            .getMajorMinecraftVersions(this.instance.id).stream().map(mv -> mv.version)
+                            .getMajorMinecraftVersions(this.instance.id).stream().map(mv -> mv.id)
                             .collect(Collectors.toList());
 
                     modrinthVersionsStream = modrinthVersionsStream.filter(